yeah

+++ pulse.cpp
@@ -3,15 +3,15 @@
 #include "pulse.h"
 
 /******************************************************************************/
-double pulse(double x, double t, double speed, double tau, double w0){
-    double t1 = t - x/speed;
-    double a = t1/tau;
+float pulse(float x, float t, float speed, float tau, float w0){
+    float t1 = t - x/speed;
+    float a = t1/tau;
     return exp(-a*a)*cos(w0*t1);
 }
 
 /******************************************************************************/
-void create_initial_dist(int Nx, double *Ey, double *Hz, double dx, double dt,
-                         double speed, int ix0, double tau, double w0,int dir){
+void create_initial_dist(int Nx, float *Ey, float *Hz, float dx, float dt,
+                         float speed, int ix0, float tau, float w0,int dir){
 
 
     for(int i=0; i<Nx; i++){
@@ -22,7 +22,7 @@ void create_initial_dist(int Nx, double *Ey, double *Hz, double dx, double dt,
 
 }
 
-void getEHL(double& E0, double& H0, int ti, int i, int ix0, double dx, double dt, double speed, double tau, double w0)
+void getEHL(float& E0, float& H0, int ti, int i, int ix0, float dx, float dt, float speed, float tau, float w0)
 {
 //    static int ti=0;
 
@@ -31,7 +31,7 @@ void getEHL(double& E0, double& H0, int ti, int i, int ix0, double dx, double dt
 }
 
 
-void getEHR(double& E0, double& H0, int ti, int i, int ix0, double dx, double dt, double speed, double tau, double w0)
+void getEHR(float& E0, float& H0, int ti, int i, int ix0, float dx, float dt, float speed, float tau, float w0)
 {
 //    static int ti=0;
 
@@ -39,7 +39,7 @@ void getEHR(double& E0, double& H0, int ti, int i, int ix0, double dx, double dt
     H0=pulse(dx*(i-0.5-ix0), (ti+0.5)*dt, speed, tau, w0);
 }
 
-void create_initial_monochrom(double *Ey, double *Hz, double dx,float k,float l,int dir)
+void create_initial_monochrom(float *Ey, float *Hz, float dx,float k,float l,int dir)
 {
     for(int i=0; (i*dx)<l; i++)
     {
